Changed UbiSense range values
Also merged X and Y constants since we need a square in order to perform
our coordinate transformations on the computer. If we wanted to specify
a non-square rectangle of limits, we would need to perform the
transformation aboard the quadrotor. We don't have enough channels for
that, and we want to minimize computational load on the onboard circuit.